refatory : [Step1] List를 일급 객체로 변경
- List<Section> sections > Sections로 변경
- 관련 테스트 코드 이관 및 로직 전체 변경
